moved post scss to own file

+++ frontend/src/components/dashboard/post_list_item.jsx
@@ -6,17 +6,24 @@ import { getOtherUserInfo } from '../../actions/user_actions';
 
 class PostListItem extends React.Component {
     componentDidMount() {
-        debugger;
         this.props.getOtherUserInfo(this.props.post.user);
+        debugger;
     }
     render() {
         const post = this.props.post;
         const editAction = this.props.editAction;
         const deleteAction = this.props.deleteAction;
-        const belongsToUser = this.props.userId === post.user;
-        debugger;
+        const belongsToUser = this.props.currentUserId === post.user;
+        const author = this.props.users.author;
+        if (!post || !author) {
+            return <div>No posts found</div>;
+        }
         return (
             <li className="post-item-container">
+                <div className="post-profile-pic">
+                    {post.user}
+                    <img src={author.image_path} alt="profile-pic" />
+                </div>
                 <Link to={`/thread/${post._id}`}>{post.text}</Link>
                 {belongsToUser && !!editAction ? <button onClick={() => editAction(post._id)}>Edit</button> : ''}
                 {belongsToUser && !!deleteAction ? <button onClick={() => deleteAction(post._id)}>Delete</button> : ''}
@@ -26,14 +33,13 @@ class PostListItem extends React.Component {
 }
 
 const mapStateToProps = (state) => {
-    debugger;
     return {
-        userId: state.session.user.id,
+        currentUserId: state.session.user.id,
+        users: state.entities.users,
     };
 };
 
 const mapDispatchToProps = (dispatch) => {
-    debugger;
     return {
         getOtherUserInfo: (userId) => dispatch(getOtherUserInfo(userId)),
     };
